requirements.txt
streamlit
together
langchain
st_copy_to_clipboard
pillow